--- app/controllers/spina/journal/primer_theme/application_controller.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Spina
-  module Journal
-    module PrimerTheme
-      # Base class from which controllers related to the journal plugin inherit
-      class ApplicationController < Spina::ApplicationController
-      end
-    end
-  end
-end
