+++ app/views/layouts/spina/conferences/primer_theme/journal/application.html.haml
