+++ app/views/layouts/spina/conferences/primer_theme/journal/issues.html.haml
@@ -1,5 +1,5 @@
 - content_for :breadcrumbs do
   = render_breadcrumbs(builder: Spina::Conferences::PrimerTheme::Breadcrumbs::Builder)
 
-= render template: 'layouts/spina/journal/primer_theme/application',
+= render template: 'layouts/spina/conferences/primer_theme/journal/application',
          locals: { author: current_account.name, description: @description, title: @title, seo_title: @title }
