+++ app/views/spina/conferences/primer_theme/journal/articles/show.html.haml
