+++ app/views/spina/conferences/primer_theme/journal/issues/_article.html.haml
