+++ app/views/spina/conferences/primer_theme/journal/issues/_issue.html.haml
