+++ app/views/spina/conferences/primer_theme/journal/issues/show.html.haml
@@ -1,7 +1,5 @@
 = render(Primer::HeadingComponent.new) do
-  = t('spina.journal.primer_theme.volume_issue',
-      volume_number: @issue.volume.number,
-      issue_number: @issue.number)
+  = t('spina.journal.primer_theme.volume_issue', volume_number: @issue.volume.number, issue_number: @issue.number)
 
 - unless @issue.title.blank?
   = render(Primer::HeadingComponent.new(tag: :h2, color: :text_secondary, mb: 2)) { @issue.title }
