+++ config/routes.rb
@@ -7,11 +7,10 @@ Spina::Engine.routes.draw do
     end
     resources :presentations, only: [:show]
     get 'cookies-info', controller: 'application'
-  end
-
-  namespace :frontend, path: 'journal', module: 'journal/primer_theme' do
-    resources :issues, only: %i[index show] do
-      resources :articles, only: %i[show]
+    scope path: 'journal', module: 'journal' do
+      resources :issues, only: %i[index show] do
+        resources :articles, only: %i[show]
+      end
     end
   end
 end
