--- test/controllers/spina/journal/primer_theme/application_controller_test.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-module Spina
-  module Journal
-    module PrimerTheme
-      class ApplicationControllerTest < ActionDispatch::IntegrationTest
-      end
-    end
-  end
-end
